Making Special:DisambiguationPageLinks work with $wgMiserMode
The default QueryPage caching assumes very simple results from the query. This
causes Special:DisambiguationPageLinks to only show partial results if it is
being fetched from querycache. To fix this, we'll need to override recache()
and fetchFromCache() in SpecialDisambiguationPageLinks.php (and also refactor a
few other things to accommodate for the limited fields available in the
querycachetwo table).

Bug: 49281
Change-Id: I830a39f760a9abbf3dbeaf43529c84f9dc21ba18