Hide disambiguator-link-added tag temporarily
This hides the tag using the same means that mw-add-media is hidden, for
instance. The tag is still 'active' and is applied to revisions, but is
hidden from tag lists including at Special:RecentChanges.

This is because currently the tag doesn't get applied to recent changes
entires, but it does for revisions which is still useful. See
T287549#7286274 for more information. The patch to solve this issue
should also revert this commit.

Bug: T287549
Change-Id: I97c22f2bf8aaa41a49954e5a15ac095bd8db491a