Make 'Review link' open WikiEditor link insertion dialog if available
When WikiEditor is not available, highlight the text as with the status
quo.

Switch ext.disambiguator to use ES6 so we can use shorthand object
property names. As a new feature, the notifications belongs as ES6-only
anyway per https://w.wiki/4D7Y

Bump minimum MW version to 1.37, as newer ES6 would cause test failures
when ran on an older branch.

Bug: T292670
Change-Id: I4447905ec2b409d5820a7ee006a2197097a99695